feat: funcionalidad de persistencia y visualizacion de datos lista

+++ src/main/java/com/solab/appdesktop/model/Catalogo.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Builder
 public class Catalogo {
     private Long id;
-    private Long numero;
+    private Integer numero;
     private String nombre;
     private String fecha;
 }
